requirements.txt
alembic==0.9.6
Babel==2.5.1
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dominate==2.3.1
elasticsearch==7.5.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.0.0
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.18.2
guess-language-spirit==0.5.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
mysqlclient==2.0.1
PyJWT==1.5.3
pyngrok==4.1.7
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.2
PyYAML==5.3.1
redis==3.2.1
requests==2.18.4
rq==1.0
six==1.11.0
SQLAlchemy==1.3.18
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
pymysql==0.10.0
cryptography==2.9.2